renamed lib.py -> json.py
