+++ python/test/test_layout.py
@@ -3,7 +3,7 @@ import sys
 from pathlib import Path
 sys.path.append(str(Path('__file__').resolve().parent))
 from source.layout import *
-from source.lib import to_json
+from source.json import to_json
 
 
 class TestLayout(unittest.TestCase):
